requirements.txt
customtkinter==5.2.0
pyinstaller==6.13.0
watchdog==3.0.0
numpy>=2.0.0
sentence-transformers>=2.2.2
huggingface-hub>=0.20.0
torch>=2.0.0
scikit-learn>=1.3.0
networkx>=3.0
faiss-cpu>=1.7.4
nltk>=3.8.1
PyMuPDF>=1.21.1
tqdm>=4.65.0
pypdf>=3.15.1